requirements.txt
Flask
python-dotenv
mcp[cli]
httpx
hashgraph-sdk

